Converted mem_ref from a pointer to an integer so it has the same value for zero as for nonzero base.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1994 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/emu-i386/simx86/syncpu.h
@@ -134,7 +134,7 @@ typedef struct {
  *	b28-31= LLRW bp#3
  */
 	unsigned int dr[8];
-	unsigned long mem_ref;
+	unsigned int mem_ref;
 /* CPU register: base(32) limit(16) */
 	DTR  GDTR;
 /* CPU register: base(32) limit(16) */
